The "Woops" Commit
Apparently I did a bunch of stuff and never committed any of it to GitHub,
so I'm doing that now.  Woops.  This works fine for Windows 7, but now
that I've upgraded to Windows 10 there's a problem:

the library used to get the screen resolution just returns 1024x720 for
some reason.  I'll probably just end up hardcoding the screen resolution
in the config file, honestly.